chore: clippy/fmt clean across workspace, gitignore Claude local settings
Drives cargo fmt --check and cargo clippy --workspace --all-targets
-D warnings to zero failures. Each fix replaces a code-quality issue with
the lint's recommended form rather than suppressing it.

Lint fixes
  - clippy::manual_pattern_char_comparison: split(['x', 'y']) form
  - clippy::manual_contains: slice.contains(&item)
  - clippy::collapsible_if: merged guard with the body condition
  - clippy::unnecessary_cast: drop redundant `as i64` on libc::SYS_*
  - clippy::field_reassign_with_default: struct literal with ..Default
  - clippy::single_match: replaced with if-let
  - clippy::manual_contains -> contains, clippy::useless_vec -> array
  - clippy::format_in_format_args: hoisted nested format! to a binding
  - clippy::manual_div_ceil-equivalent: is_multiple_of
  - clippy::err_expect: result.unwrap_err()
  - clippy::redundant_clone: std::slice::from_ref
  - unreachable_code: unix-only tests gated with #[cfg(unix)] instead of
    early-return after a #[cfg(not(unix))] block

Structural fix: tests-after-items in src/proxy.rs
  CONNECT tunnel implementation (handle_connect, handle_connect_inner,
  blind_relay) lived after the #[cfg(test)] mod tests block, tripping
  clippy::items_after_test_module. Moved tests to the end of the file
  so production code is contiguous and tests are unambiguously last.

Structural fix: wasm benches behind feature flag
  benches/pipeline.rs unconditionally imported gvm_proxy::wasm_engine and
  gvm_engine, both gated on the `wasm` cargo feature. Split bench
  registration into native_benches (always) and wasm_benches (cfg
  feature = "wasm"), with two criterion_main! arms. Bench binary now
  builds on default features.

AppState test fixtures
  Tests/integration.rs constructed AppState directly in 10 places. The
  ExitReason work added start_time, request_counter, and ca_expires_days
  fields. Added the three fields to every fixture (Instant::now,
  AtomicU64::new(0), None).

Filesystem unix-only tests
  symlink_escape_is_discarded and execute_merge_rejects_symlinks were
  Linux-only via early-return after #[cfg(not(unix))]. Replaced with
  #[cfg(unix)] on the test attribute itself — eliminates the
  unreachable_code warning and makes the test/platform mapping explicit.

gitignore
  .claude/settings.local.json — local Claude Code agent settings,
  per-machine and not part of the repo.

+++ crates/gvm-sandbox/src/seccomp.rs
@@ -237,7 +237,7 @@ fn build_default_filter(default_action: SeccompAction) -> Result<seccompiler::Bp
 pub fn allowed_syscall_count() -> usize {
     let mut allow_rules: BTreeMap<i64, Vec<SeccompRule>> = BTreeMap::new();
     insert_base_syscalls(&mut allow_rules);
-    allow_rules.insert(libc::SYS_socket as i64, vec![]);
+    allow_rules.insert(libc::SYS_socket, vec![]);
     for sys in [
         libc::SYS_connect,
         libc::SYS_socketpair,
@@ -257,7 +257,7 @@ pub fn allowed_syscall_count() -> usize {
         libc::SYS_accept,
         libc::SYS_accept4,
     ] {
-        allow_rules.insert(sys as i64, vec![]);
+        allow_rules.insert(sys, vec![]);
     }
     allow_rules.len()
 }
